submission/requirements.txt
# Core ML/AI Libraries (for local model)
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# ML Utilities
scikit-learn>=1.3.0

# Configuration
PyYAML>=6.0

# Explainability
shap>=0.41.0

# API Support
requests>=2.31.0
openai>=1.0.0  # Optional: for OpenAI API usage

# Utilities
tqdm>=4.65.0
